progress: 41.58% of code matched
Source-commit: c6ca4f09f3553ef7c2c7804b4c0055cb56458af1

+++ src/game/game_fn_801E3388.c
@@ -98,7 +98,7 @@ void fn_801E3388(Object* object)
                                                 output_a, output_d,
                                                 resource, lbl_80651248));
             *output_c = *output_b;
-            if ((object->flags & 0xF) == 8 && resource != 0) {
+            if ((object->flags & 0xF) == 8 && (int)resource != 0) {
                 void* second_b;
                 void** second_d;
                 u8 kind;
